+++ src/Constants.svelte
@@ -1,0 +1,15 @@
+<script context="module">
+    
+    export let width = window.innerWidth;
+    export let height = window.innerWidth/1920*1080;
+    export  const numBarShow = 3;
+    export const startingPoint = width/2.5
+    export const layerWidth = width-startingPoint;
+    export const lineWidth = height/700;
+    export const layerInstLineWidth = height/15
+    export const maxAmpRadius = height/8
+
+
+
+
+</script>
